Rename exit button file and clean up edge case
